requirements.txt
openai-whisper
pyannote.audio
ffmpeg-python
torch
